requirements.txt
python-dotenv
ipykernel
pandas
scikit-learn
matplotlib
openai==0.28